qa: resolve Psalm errors
With changes to laminas-db, a number of types are now known, which allows less strict checking in laminas-feed.
However, there are also a few cases now where we have "new" `MixedAssignment` violations, which are perfectly reasonable as we're working with associative data structures.
I've added these three new violations to the baseline.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>