fix: ensure PuSH callback subscriber tests are using valid collaborators
PHPUnit mocks will return null for method calls where no expectations were provided.
As a result, a mock that was expected to return a `Laminas\Db\ResultSet\ResultSetInterface` was returning `null`.
Previously, we would do `if ($result && count($result))`, but because laminas-db is now documented as returning a `ResultSetInterface` for the method in question, we dropped the check to see if `$result` was truthy.
This then results in an error due to calling `count()` on a non-countable value (`null`).

The fix was to update the test setup to have the method in question return a `ResultSetInterface` stub.

In doing so, I discovered two test methods named differently, but with the exact same expectations.
Since it was unclear what the second was doing, I marked it skipped and added a TODO for it.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>