Mention creation of Feeds in composer.json
This commit adds a mention that the package can also be used to create - and not only consume - feeds.

It also adds the "rss" and the "atom" keywords. Even though those two phrases are already part of the description it can not hurt to have them in the keywords as well.

Signed-off-by: Andreas Heigl <heiglandreas@users.noreply.github.com>